requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.0
pydantic-settings==2.6.1
openai==1.54.0
anthropic==0.39.0
google-generativeai==0.8.3
tiktoken==0.8.0
httpx==0.27.2
